+++ apps/shelf/funcs/echo.c
@@ -17,25 +17,12 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 -----------------------------------------------------------------------
 */
 
-#include "farix.h"
+#include <stdio.h>
 
-#define SYS_WRITE 1
+#include "cmds.h"
 
-// Took from the journal (finally became useful)
-int main(int argc, char** argv) {
-    const char* msg = "Farix is alive!\n";
-    int len = 16;
+void cmd_echo(const char* args) {
+    if (!args) return;
 
-    asm volatile (
-        "mov %0, %%eax\n"  // SYS_WRITE
-        "mov %1, %%ebx\n"  // file descriptor 1 (stdout)
-        "mov %2, %%ecx\n"  // buffer pointer
-        "mov %3, %%edx\n"  // length
-        "int $0x80\n"
-        :
-        : "i"(SYS_WRITE), "i"(1), "r"(msg), "r"(len)
-        : "eax", "ebx", "ecx", "edx"
-    );
-
-    while(1);
+    printf("%s\n", args);
 }